requirements.txt
streamlit>=1.30.0
openai>=1.12.0
pandas>=2.1.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
google-search-results>=2.4.0
thefuzz>=0.22.0
python-Levenshtein>=0.25.0
